fix(lint): resolve 17 pre-existing ruff check violations
Backend lint has been red on every PR for some time, masking format
failures behind it. This commit clears the ruff check errors so the
job can progress to the format step (handled in the next commit).

Changes:
- auth.py: introduce _fire_and_forget helper that retains a strong
  reference to fire-and-forget asyncio tasks (RUF006). Three signup-
  notification call sites now route through it. Behavior unchanged —
  the tasks still don't block the request, but they no longer risk
  GC mid-flight.
- billing.py: drop unused Subscription import (F401).
- analytics.py, str_calc.py, property_service.py: replace ambiguous
  U+00D7 (×) with ASCII 'x' in docstrings/strings (RUF001/RUF002).
- auth_service.py: drop now-redundant string-quoted User annotation;
  __future__ annotations is active and User is imported at module
  level (UP037).
- iq_verdict_service.py: drop redundant int() around round() — Python's
  round() returns int when no ndigits is provided (RUF046).
- zillow_client.py: drop dead is_fsba assignment; FSBA is the default
  seller_type (line 727), no flag needed (F841).
- I001 import-sort fixes applied via ruff --fix.

+++ backend/app/services/iq_verdict_service.py
@@ -545,7 +545,7 @@ def _get_cumulative_investor_pct(deal_gap_pct: float, region_key: str) -> int:
             total += pct * (hi - deal_gap_pct) / (hi - lo)
         else:
             total += pct
-    rounded = int(round(total))
+    rounded = round(total)
     return max(1, min(99, rounded))
